requirements.txt
pandas==1.5.3
tabula-py==2.8.0
openpyxl==3.0.10
xmltodict==0.13.0
python-dotenv==1.0.0
sqlalchemy==1.4.46
psycopg2-binary==2.9.6
pika==1.3.1
pydantic==1.10.7